adding boolean methods file
